+++ gaios/Extensions/UIImage.swift
@@ -1,0 +1,19 @@
+import AVFoundation
+import UIKit
+public extension UIImage {
+    func resize(_ width: Int, _ height: Int) -> UIImage {
+        let maxSize = CGSize(width: width, height: height)
+        let availableRect = AVFoundation.AVMakeRect(
+            aspectRatio: self.size,
+            insideRect: .init(origin: .zero, size: maxSize)
+        )
+        let targetSize = availableRect.size
+        let format = UIGraphicsImageRendererFormat()
+        format.scale = 1
+        let renderer = UIGraphicsImageRenderer(size: targetSize, format: format)
+        let resized = renderer.image { _ in
+            self.draw(in: CGRect(origin: .zero, size: targetSize))
+        }
+        return resized
+    }
+}