+++ gaios/Wallet/WalletViewController/Cells/PromoCell/PromoLayout2Cell.swift
@@ -3,19 +3,16 @@ import UIKit
 import gdk
 import core
 
-enum PromoType {
-    case jadeDiscount
-}
-
-class PromoCell: UITableViewCell {
+class PromoLayout2Cell: UITableViewCell {
 
     @IBOutlet weak var bg: UIView!
+    @IBOutlet weak var lblInfo: UILabel!
     @IBOutlet weak var lblTitle: UILabel!
     @IBOutlet weak var lblHint: UILabel!
     @IBOutlet weak var btnAction: UIButton!
     @IBOutlet weak var icon: UIImageView!
     @IBOutlet weak var btnDismiss: UIButton!
-    @IBOutlet weak var iconWrap: UIView!
+    @IBOutlet weak var gradient: UIView!
 
     var model: PromoCellModel?
     var onAction: (() -> Void)?
@@ -44,33 +41,47 @@ class PromoCell: UITableViewCell {
         self.onAction = onAction
         self.onDismiss = onDismiss
 
-        lblTitle.text = model.promo.title_small?.htmlDecoded
-        lblHint.text = model.promo.text_small?.htmlDecoded
-        btnAction.setTitle(model.promo.cta_small?.htmlDecoded, for: .normal)
+        lblInfo.text = model.promo.overline_small
+        lblTitle.text = model.promo.title_small
+        lblHint.text = model.promo.text_small
+        btnAction.setTitle(model.promo.cta_small, for: .normal)
 
+        self.icon.image = model.promo.thumb
         if let thumb = model.promo.thumb {
             let w = thumb.size.width
             let h = thumb.size.height
-            if h > 0 {
-                let nW = 100 * w / h
-                let resized = thumb.resize(Int(nW), 100)
-                iconWrap.isHidden = false
+            if w > 0 {
+                let nW = (UIScreen.main.bounds.width - 50)
+                let nH = nW * h / w
+                let resized = thumb.resize(Int(nW), Int(nH))
                 icon.image = resized
             }
-        } else {
-            iconWrap.isHidden = true
         }
+        icon.clipsToBounds = true
     }
 
     func setStyle() {
         bg.layer.cornerRadius = 6.0
         bg.borderWidth = 1.0
         bg.borderColor = .white.withAlphaComponent(0.1)
-        lblTitle.setStyle(.subTitle)
-        lblTitle.font = UIFont.systemFont(ofSize: 20.0, weight: .semibold)
-        lblHint.setStyle(.txtCard)
+        lblInfo.setStyle(.txtSmaller)
+        lblTitle.setStyle(.subTitle24)
+        lblHint.setStyle(.txtSmaller)
         btnAction.setStyle(.primary)
         btnDismiss.setImage(UIImage(named: "cancel")?.maskWithColor(color: .white), for: .normal)
+    }
+
+    override func draw(_ rect: CGRect) {
+        super.draw(rect)
+
+//        DispatchQueue.main.asyncAfter(deadline: DispatchTime.now() + 0.1) {
+//            let gLayer = self.makeGradient(colours: [.clear, .black.withAlphaComponent(0.5)], locations: [0.5, 1.0])
+//            gLayer.startPoint = CGPoint(x: 0.0, y: 0.0)
+//            gLayer.endPoint = CGPoint(x: 0.0, y: 1.0)
+//            gLayer.frame = self.gradient.bounds
+//            gLayer.masksToBounds = true
+//            self.gradient.layer.addSublayer(gLayer)
+//        }
     }
 
     @IBAction func btnAction(_ sender: Any) {
@@ -82,24 +93,5 @@ class PromoCell: UITableViewCell {
             PromoManager.shared.onDismiss(promo: promo, source: source)
         }
         onDismiss?()
-    }
-}
-
-public extension UIImage {
-
-    func resize(_ width: Int, _ height: Int) -> UIImage {
-        let maxSize = CGSize(width: width, height: height)
-        let availableRect = AVFoundation.AVMakeRect(
-            aspectRatio: self.size,
-            insideRect: .init(origin: .zero, size: maxSize)
-        )
-        let targetSize = availableRect.size
-        let format = UIGraphicsImageRendererFormat()
-        format.scale = 1
-        let renderer = UIGraphicsImageRenderer(size: targetSize, format: format)
-        let resized = renderer.image { _ in
-            self.draw(in: CGRect(origin: .zero, size: targetSize))
-        }
-        return resized
     }
 }