+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 - Update gdk to 0.74.0
+- Added support for multilayout and video promos
 
 ## [4.1.0] - 2024-11-20